refactor(web): move history import conversion to api
